doctor specialities are up to date

+++ project/src/main/java/regenaration/team4/repositories/AppointmentRepository.java
@@ -3,5 +3,5 @@ package regenaration.team4.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import regenaration.team4.entities.Appointment;
 
-public interface AppointmentRepository extends JpaRepository<Appointment,Integer> {
+public interface AppointmentRepository extends JpaRepository<Appointment,Long> {
 }
